+++ asn1/src/asym_key_dilithium_ed448.c
@@ -101,9 +101,11 @@ int x509_mldsa_ed448_private_key(void *context, size_t hdrlen,
 				 unsigned char tag, const uint8_t *value,
 				 size_t vlen)
 {
+	struct workspace {
+		struct lc_dilithium_pk pk;
+		struct lc_dilithium_sk sk;
+	};
 	struct lc_x509_key_data *keys = context;
-	struct lc_dilithium_pk pk;
-	struct lc_dilithium_sk sk;
 	struct lc_dilithium_ed448_sk *dilithium_sk =
 		keys->sk.dilithium_ed448_sk;
 	const uint8_t *data, *ed448_src_key;
@@ -111,6 +113,7 @@ int x509_mldsa_ed448_private_key(void *context, size_t hdrlen,
 	size_t datalen, dilithium_src_key_len, ed448_src_key_len;
 	enum lc_dilithium_type dilithium_type = LC_DILITHIUM_UNKNOWN;
 	int ret;
+	LC_DECLARE_MEM(ws, struct workspace, sizeof(uint64_t));
 
 	(void)hdrlen;
 	(void)tag;
@@ -173,9 +176,10 @@ int x509_mldsa_ed448_private_key(void *context, size_t hdrlen,
 	 * First the ML-DSA seed, then the traditional SK.
 	 */
 	CKINT(lc_dilithium_keypair_from_seed(
-		&pk, &sk, data, LC_X509_PQC_SK_SEED_SIZE, dilithium_type));
+		&ws->pk, &ws->sk, data, LC_X509_PQC_SK_SEED_SIZE,
+		dilithium_type));
 	CKINT(lc_dilithium_sk_ptr(&dilithium_src_key, &dilithium_src_key_len,
-				  &sk));
+				  &ws->sk));
 
 	ed448_src_key = data + LC_X509_PQC_SK_SEED_SIZE;
 	ed448_src_key_len = LC_ED448_SECRETKEYBYTES;
@@ -187,8 +191,7 @@ int x509_mldsa_ed448_private_key(void *context, size_t hdrlen,
 	printf_debug("Loaded composite public key of size %zu\n", datalen);
 
 out:
-	lc_memset_secure(&pk, 0, sizeof(pk));
-	lc_memset_secure(&sk, 0, sizeof(sk));
+	LC_RELEASE_MEM(ws);
 	return ret;
 }