+++ asn1/src/pkcs7_parser.c
@@ -42,8 +42,8 @@
  * ASN.1 parser support functions
  ******************************************************************************/
 
-static __always_inline int test_and_set_bit(unsigned long nr,
-					    volatile unsigned long *addr)
+static __always_inline int lc_test_and_set_bit(unsigned long nr,
+					       volatile unsigned long *addr)
 {
 	unsigned long *p = ((unsigned long *)addr);
 	unsigned long old = *p;
@@ -512,7 +512,7 @@ int pkcs7_sig_note_authenticated_attr(void *context, size_t hdrlen,
 #pragma GCC diagnostic ignored "-Wswitch-enum"
 	switch (ctx->last_oid) {
 	case OID_contentType:
-		if (test_and_set_bit(sinfo_has_content_type, &sinfo->aa_set))
+		if (lc_test_and_set_bit(sinfo_has_content_type, &sinfo->aa_set))
 			goto repeated;
 		content_type = look_up_OID(value, vlen);
 		if (content_type != ctx->msg->data_type) {
@@ -525,7 +525,7 @@ int pkcs7_sig_note_authenticated_attr(void *context, size_t hdrlen,
 		return 0;
 
 	case OID_signingTime:
-		if (test_and_set_bit(sinfo_has_signing_time, &sinfo->aa_set))
+		if (lc_test_and_set_bit(sinfo_has_signing_time, &sinfo->aa_set))
 			goto repeated;
 		/* Should we check that the signing time is consistent
 		 * with the signer's X.509 cert?
@@ -534,7 +534,8 @@ int pkcs7_sig_note_authenticated_attr(void *context, size_t hdrlen,
 					value, vlen);
 
 	case OID_messageDigest:
-		if (test_and_set_bit(sinfo_has_message_digest, &sinfo->aa_set))
+		if (lc_test_and_set_bit(sinfo_has_message_digest,
+					&sinfo->aa_set))
 			goto repeated;
 		if (tag != ASN1_OTS)
 			return -EBADMSG;
@@ -543,7 +544,7 @@ int pkcs7_sig_note_authenticated_attr(void *context, size_t hdrlen,
 		return 0;
 
 	case OID_smimeCapabilites:
-		if (test_and_set_bit(sinfo_has_smime_caps, &sinfo->aa_set))
+		if (lc_test_and_set_bit(sinfo_has_smime_caps, &sinfo->aa_set))
 			goto repeated;
 		if (ctx->msg->data_type != OID_msIndirectData) {
 			printf_debug(
@@ -560,12 +561,12 @@ int pkcs7_sig_note_authenticated_attr(void *context, size_t hdrlen,
 		 * are also used as extendedKeyUsage types in X.509 certs.
 		 */
 	case OID_msSpOpusInfo:
-		if (test_and_set_bit(sinfo_has_ms_opus_info, &sinfo->aa_set))
+		if (lc_test_and_set_bit(sinfo_has_ms_opus_info, &sinfo->aa_set))
 			goto repeated;
 		goto authenticode_check;
 	case OID_msStatementType:
-		if (test_and_set_bit(sinfo_has_ms_statement_type,
-				     &sinfo->aa_set))
+		if (lc_test_and_set_bit(sinfo_has_ms_statement_type,
+					&sinfo->aa_set))
 			goto repeated;
 	authenticode_check:
 		if (ctx->msg->data_type != OID_msIndirectData) {